requirements.txt
Flask
flask_login
SQLAlchemy
psycopg2-binary
python-dotenv
python-docx